Added ParseCgiResponse.cpp(just started)

+++ srcs/Client.cpp
@@ -346,7 +346,7 @@ ssize_t     Client::getChunkData(){
     size_t strSize = _rawBuffer.find("\r\n");
     if (strSize == std::string::npos)
         return 0;
-    if (strSize != _chunkSize)
+    if (static_cast<ssize_t>(strSize) != _chunkSize)
         return -1;
     std::string chunkStr = _rawBuffer.substr(0, strSize);
     store(chunkStr, REQUEST);
